checks if person is present

+++ lib/tasks/remove_roles.rake
@@ -2,20 +2,24 @@ namespace :role do
   desc "remove assister and cac roles from users"
   # Usage rake role:remove
   task :remove => [:environment] do
-     
+
     CSV.foreach("bad_users.csv") do |row|
       email = row[2]
       role = row[3]
       hbx = row[4]
-      
+
       if hbx.present?
         person = Person.where(hbx_id:hbx).first
-        if role == "assister" && person.assister_role.present?
-          person.assister_role.destroy
-          person.update(is_disabled:true)
-        elsif role == "cac" && person.csr_role.present?
-          person.csr_role.destroy
-          person.update(is_disabled:true)
+        if person.present?
+          if role == "assister" && person.assister_role.present?
+            person.assister_role.destroy
+            person.update(is_disabled:true)
+          elsif role == "cac" && person.csr_role.present?
+            person.csr_role.destroy
+            person.update(is_disabled:true)
+          end
+        else
+          puts "No person found for HBX ID: #{hbx}"
         end
       elsif email.present?
         user = User.where(email:email).first
@@ -29,9 +33,8 @@ namespace :role do
           end
         end
       end
-      
+
     end
     puts "Successfully removed roles"
   end
 end
-
